backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sympy>=1.12
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
qdrant-client>=1.7.1
sentence-transformers>=2.2.0
tavily-python>=0.3.0
requests>=2.31.0
ollama>=0.1.7
